+++ LibTest/ffi/Short/Short_A01_t02.dart
@@ -20,7 +20,7 @@ import "../../../Utils/expect.dart";
 
 void main() {
   Pointer<Short> p1 = calloc<Short>();
-  Pointer<Int32> p2 = new Pointer<Int32>.fromAddress(p1.address);
+  Pointer<Uint16> p2 = new Pointer<Uint16>.fromAddress(p1.address);
   try {
     p2.value = 42;
     Expect.equals(42, p1.value);
